udpated files

+++ backend/app/models/candle.py
@@ -12,4 +12,4 @@ class Candle(Base):
     high = Column(Float)
     low = Column(Float)
     close = Column(Float)
-    volume = Column(BigInteger)
+    volume = Column(Float)
